refactor: split PlayerOverlay into player/ directory (#15)
Split the 804-line PlayerOverlay.tsx into focused modules:
- player/PlayerOverlay.tsx — Main overlay with navigation
- player/VideoSlide.tsx — Individual video player + controls
- player/ShareSheet.tsx — Share bottom sheet with TikTokIcon
- player/QuestionOverlay.tsx — Question text overlay
- player/types.ts — Shared prop interfaces

Old components/PlayerOverlay.tsx kept as re-export for safety.
